Fix average lands to cast numerator and denominator as decimal to properly get quotient as decimal result.

+++ ScryfallConnector/DeckStatisticsForm.cs
@@ -286,7 +286,7 @@ namespace ScryfallConnector.Classes
                             break;
                     }
                 }
-                averageLandsPerHand = totalLands / counter;
+                averageLandsPerHand = (decimal)totalLands / (decimal)counter;
                 averageLandsPerHand = Math.Round(averageLandsPerHand, 2);
                 if (!verboseLog)
                 {
